+++ docs/conf.py
@@ -7,7 +7,7 @@ from datetime import datetime
 project = "LessonKit"
 copyright = f"{datetime.now().year}, LessonKit contributors"
 author = "LessonKit contributors"
-release = "1.2.0"
+release = "1.3.0"
 
 extensions = [
     "myst_parser",